Move Parallax to extension method
